# openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/api/v1/users/profile": {"get": {"tags": ["Users"], "summary": "Get Profile User", "operationId": "get_profile_user_api_v1_users_profile_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GetProfile"}}}}}}}, "/api/v1/users/profile/create": {"post": {"tags": ["Users"], "summary": "Create Profile User", "operationId": "create_profile_user_api_v1_users_profile_create_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateProfile"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OkResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/users/profile/update": {"put": {"tags": ["Users"], "summary": "Update Profile User", "operationId": "update_profile_user_api_v1_users_profile_update_put", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateProfile"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OkResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/login": {"post": {"tags": ["Auth"], "summary": "Login User", "operationId": "login_user_api_v1_auth_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginUserSchema"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OkResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/refresh_token": {"post": {"tags": ["Auth"], "summary": "Auth Refresh Jwt", "operationId": "auth_refresh_jwt_api_v1_auth_refresh_token_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OkResponse"}}}}}}}, "/api/v1/auth/check_auth": {"get": {"tags": ["Auth"], "summary": "Check Auth User And Get Datauser", "operationId": "check_auth_user_and_get_datauser_api_v1_auth_check_auth_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CheckAuth"}}}}}}}, "/api/v1/auth/logout": {"post": {"tags": ["Auth"], "summary": "Auth Refresh Jwt", "operationId": "auth_refresh_jwt_api_v1_auth_logout_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/tests/{test_id}": {"get": {"tags": ["Random tests", "Static tests"], "summary": "Get Test", "operationId": "get_test_api_v1_tests__test_id__get", "parameters": [{"name": "test_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Test Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/tests/{test_id}/start": {"get": {"tags": ["Random tests"], "summary": "Generate Questions", "operationId": "generate_questions_api_v1_tests__test_id__start_get", "parameters": [{"name": "test_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Test Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StartFinishTestSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/tests/{test_id}/question/{q_num}": {"get": {"tags": ["Random tests"], "summary": "Get Question Answers", "operationId": "get_question_answers_api_v1_tests__test_id__question__q_num__get", "parameters": [{"name": "test_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Test Id"}}, {"name": "q_num", "in": "path", "required": true, "schema": {"type": "integer", "title": "Q Num"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GetQuestionAndAnswersSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["Random tests"], "summary": "Add Answer", "operationId": "add_answer_api_v1_tests__test_id__question__q_num__post", "parameters": [{"name": "test_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Test Id"}}, {"name": "q_num", "in": "path", "required": true, "schema": {"type": "integer", "title": "Q Num"}}, {"name": "answer_id", "in": "query", "required": true, "schema": {"type": "integer", "title": "Answer Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OkStatusSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/tests/{test_id}/finish": {"post": {"tags": ["Random tests"], "summary": "Add Results Test", "operationId": "add_results_test_api_v1_tests__test_id__finish_post", "parameters": [{"name": "test_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Test Id"}}, {"name": "time_execution", "in": "query", "required": true, "schema": {"type": "integer", "title": "Time Execution"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CheckNewRewardSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/tests/{test_id}/result": {"get": {"tags": ["Random tests"], "summary": "Get Result Test", "operationId": "get_result_test_api_v1_tests__test_id__result_get", "parameters": [{"name": "test_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Test Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/src__api_v1__tests__schemas__ResultTestSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/static_test/{test_id}/start": {"get": {"tags": ["Static tests"], "summary": "Get Questions", "operationId": "get_questions_api_v1_static_test__test_id__start_get", "parameters": [{"name": "test_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Test Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/static_test/{test_id}/question/{q_num}": {"get": {"tags": ["Static tests"], "summary": "Get Question Answers", "operationId": "get_question_answers_api_v1_static_test__test_id__question__q_num__get", "parameters": [{"name": "test_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Test Id"}}, {"name": "q_num", "in": "path", "required": true, "schema": {"type": "integer", "title": "Q Num"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["Static tests"], "summary": "Add Answer", "operationId": "add_answer_api_v1_static_test__test_id__question__q_num__post", "parameters": [{"name": "test_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Test Id"}}, {"name": "q_num", "in": "path", "required": true, "schema": {"type": "integer", "title": "Q Num"}}, {"name": "answer_id", "in": "query", "required": true, "schema": {"type": "integer", "title": "Answer Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OkStatus"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/static_test/{test_id}/finish": {"post": {"tags": ["Static tests"], "summary": "Finish Test", "operationId": "finish_test_api_v1_static_test__test_id__finish_post", "parameters": [{"name": "test_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Test Id"}}, {"name": "time_execution", "in": "query", "required": true, "schema": {"type": "integer", "title": "Time Execution"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TestPassed"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/static_test/{test_id}/result": {"get": {"tags": ["Static tests"], "summary": "Result Test", "operationId": "result_test_api_v1_static_test__test_id__result_get", "parameters": [{"name": "test_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Test Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/src__api_v1__static_test__schemas__ResultTestSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/topics": {"get": {"tags": ["Topics and Sections for main"], "summary": "Get Topics", "operationId": "get_topics_api_v1_topics_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/sections_topics": {"get": {"tags": ["Topics and Sections for main"], "summary": "Get Sections Topics", "operationId": "get_sections_topics_api_v1_sections_topics_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/section/{section_topic_id}": {"get": {"tags": ["Topics and Sections for main"], "summary": "Tests Case In Section", "operationId": "tests_case_in_section_api_v1_section__section_topic_id__get", "parameters": [{"name": "section_topic_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Section Topic Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/section/upload_image/{section_topic_id}": {"post": {"tags": ["Topics and Sections for main"], "summary": "Upload Image Section", "operationId": "upload_image_section_api_v1_section_upload_image__section_topic_id__post", "parameters": [{"name": "section_topic_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Section Topic Id"}}], "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_upload_image_section_api_v1_section_upload_image__section_topic_id__post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/story/{story_id}": {"post": {"tags": ["Story"], "summary": "Upload Image", "operationId": "upload_image_api_v1_story__story_id__post", "parameters": [{"name": "story_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Story Id"}}], "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_upload_image_api_v1_story__story_id__post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/story/": {"get": {"tags": ["Story"], "summary": "Get Stories", "operationId": "get_stories_api_v1_story__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/GetStories"}, "type": "array", "title": "Response Get Stories Api V1 Story  Get"}}}}}}}, "/api/v1/rewards/": {"get": {"tags": ["Rewards"], "summary": "Get All", "operationId": "get_all_api_v1_rewards__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/rewards/user": {"get": {"tags": ["Rewards"], "summary": "Get All User Rewards", "operationId": "get_all_user_rewards_api_v1_rewards_user_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GetAllUserRewards"}}}}}}}, "/api/v1/cases/{case_id}": {"get": {"tags": ["Cases"], "summary": "Get Case", "operationId": "get_case_api_v1_cases__case_id__get", "parameters": [{"name": "case_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Case Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/cases/{case_id}/start": {"get": {"tags": ["Cases"], "summary": "Start Case", "operationId": "start_case_api_v1_cases__case_id__start_get", "parameters": [{"name": "case_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Case Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/cases/{case_id}/finish": {"post": {"tags": ["Cases"], "summary": "Finish Case", "operationId": "finish_case_api_v1_cases__case_id__finish_post", "parameters": [{"name": "case_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Case Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OkResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/leaderboard/": {"get": {"tags": ["Leaderboard"], "summary": "Get Leaderboard", "operationId": "get_leaderboard_api_v1_leaderboard__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LeaderBoardResponseSchemas"}}}}}}}}, "components": {"schemas": {"AnswerSchema": {"properties": {"id": {"type": "integer", "title": "Id"}, "answer_text": {"type": "string", "title": "Answer Text"}}, "type": "object", "required": ["id", "answer_text"], "title": "AnswerSchema"}, "Body_upload_image_api_v1_story__story_id__post": {"properties": {"file": {"type": "string", "format": "binary", "title": "File"}}, "type": "object", "required": ["file"], "title": "Body_upload_image_api_v1_story__story_id__post"}, "Body_upload_image_section_api_v1_section_upload_image__section_topic_id__post": {"properties": {"file": {"type": "string", "format": "binary", "title": "File"}}, "type": "object", "required": ["file"], "title": "Body_upload_image_section_api_v1_section_upload_image__section_topic_id__post"}, "CheckAuth": {"properties": {"is_authenticated": {"type": "boolean", "title": "Is Authenticated"}, "user_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "User Id"}, "first_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "First Name"}, "last_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Last Name"}, "middle_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Middle Name"}}, "type": "object", "required": ["is_authenticated"], "title": "CheckAuth"}, "CheckNewRewardSchema": {"properties": {"ok": {"type": "boolean", "title": "Ok"}, "message": {"type": "string", "title": "Message"}, "reward_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Reward Id"}}, "type": "object", "required": ["ok", "message", "reward_id"], "title": "CheckNewRewardSchema"}, "Color": {"properties": {"top": {"type": "string", "title": "Top"}, "bottom": {"type": "string", "title": "Bottom"}}, "type": "object", "required": ["top", "bottom"], "title": "Color"}, "CorrectAnswerSchema": {"properties": {"id": {"type": "integer", "title": "Id"}, "correct_answer_text": {"type": "string", "title": "Correct Answer Text"}}, "type": "object", "required": ["id", "correct_answer_text"], "title": "CorrectAnswerSchema"}, "CreateProfile": {"properties": {"first_name": {"type": "string", "title": "First Name"}, "last_name": {"type": "string", "title": "Last Name"}, "middle_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Middle Name"}}, "type": "object", "required": ["first_name", "last_name"], "title": "CreateProfile"}, "GetAllUserRewards": {"properties": {"id": {"type": "integer", "title": "Id"}, "reward": {"type": "string", "title": "Reward"}, "description": {"type": "string", "title": "Description"}, "image_url": {"type": "string", "title": "Image Url"}, "topic_id": {"type": "integer", "title": "Topic Id"}, "topic": {"type": "string", "title": "Topic"}}, "type": "object", "required": ["id", "reward", "description", "image_url", "topic_id", "topic"], "title": "GetAllUserRewards"}, "GetProfile": {"properties": {"first_name": {"type": "string", "title": "First Name"}, "last_name": {"type": "string", "title": "Last Name"}, "middle_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Middle Name"}, "bio": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Bio"}, "id": {"type": "integer", "title": "Id"}, "history_test": {"items": {"$ref": "#/components/schemas/HistoryTestSchema"}, "type": "array", "title": "History Test"}}, "type": "object", "required": ["first_name", "last_name", "id", "history_test"], "title": "GetProfile"}, "GetQuestionAndAnswersSchema": {"properties": {"q_num": {"type": "integer", "title": "Q Num"}, "question_text": {"type": "string", "title": "Question Text"}, "answers": {"items": {"$ref": "#/components/schemas/AnswerSchema"}, "type": "array", "title": "Answers"}, "tip": {"type": "string", "title": "Tip"}}, "type": "object", "required": ["q_num", "question_text", "answers", "tip"], "title": "GetQuestionAndAnswersSchema"}, "GetStories": {"properties": {"id": {"type": "integer", "title": "Id"}, "title": {"type": "string", "title": "Title"}, "body": {"items": {}, "type": "array", "title": "Body"}, "img_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Img Url"}, "position": {"$ref": "#/components/schemas/Position"}, "size": {"$ref": "#/components/schemas/Size"}, "color": {"$ref": "#/components/schemas/Color"}}, "type": "object", "required": ["id", "title", "body", "img_url", "position", "size", "color"], "title": "GetStories"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "HistoryTestSchema": {"properties": {"id": {"type": "integer", "title": "Id"}, "title": {"type": "string", "title": "Title"}, "time_execution": {"type": "integer", "title": "Time Execution"}, "score": {"type": "integer", "title": "Score"}}, "type": "object", "required": ["id", "title", "time_execution", "score"], "title": "HistoryTestSchema"}, "LeaderBoardResponseSchemas": {"properties": {"top_users": {"items": {"$ref": "#/components/schemas/LeaderBoardSchema"}, "type": "array", "title": "Top Users"}, "current_user": {"$ref": "#/components/schemas/LeaderBoardSchema"}}, "type": "object", "required": ["top_users", "current_user"], "title": "LeaderBoardResponseSchemas"}, "LeaderBoardSchema": {"properties": {"place": {"type": "integer", "title": "Place"}, "user_id": {"type": "integer", "title": "User Id"}, "username": {"type": "string", "title": "Username"}, "score": {"type": "integer", "title": "Score"}}, "type": "object", "required": ["place", "user_id", "username", "score"], "title": "LeaderBoardSchema"}, "LoginUserSchema": {"properties": {"username": {"type": "string", "maxLength": 40, "minLength": 1, "title": "Username"}, "password": {"type": "string", "minLength": 3, "title": "Password"}}, "type": "object", "required": ["username", "password"], "title": "LoginUserSchema"}, "OkResponse": {"properties": {"ok": {"type": "boolean", "title": "Ok"}, "message": {"type": "string", "title": "Message"}}, "type": "object", "required": ["ok", "message"], "title": "OkResponse"}, "OkStatus": {"properties": {"ok": {"type": "boolean", "title": "Ok"}, "message": {"type": "string", "title": "Message"}}, "type": "object", "required": ["ok", "message"], "title": "OkStatus"}, "OkStatusSchema": {"properties": {"ok": {"type": "boolean", "title": "Ok"}, "message": {"type": "string", "title": "Message"}}, "type": "object", "required": ["ok", "message"], "title": "OkStatusSchema"}, "Position": {"properties": {"bottom": {"type": "string", "title": "Bottom"}, "right": {"type": "string", "title": "Right"}}, "type": "object", "required": ["bottom", "right"], "title": "Position"}, "ResultAnswersUserSchema": {"properties": {"q_num": {"type": "integer", "title": "Q Num"}, "question_text": {"type": "string", "title": "Question Text"}, "user_answer": {"$ref": "#/components/schemas/UserAnswerSchema"}, "is_correct": {"type": "boolean", "title": "Is Correct"}, "correct_answer": {"$ref": "#/components/schemas/CorrectAnswerSchema"}}, "type": "object", "required": ["q_num", "question_text", "user_answer", "is_correct", "correct_answer"], "title": "ResultAnswersUserSchema"}, "Size": {"properties": {"width": {"type": "integer", "title": "Width"}, "height": {"type": "integer", "title": "Height"}}, "type": "object", "required": ["width", "height"], "title": "Size"}, "StartFinishTestSchema": {"properties": {"ok": {"type": "boolean", "title": "Ok"}, "test_was_passed": {"type": "boolean", "title": "Test Was Passed"}, "message": {"type": "string", "title": "Message"}}, "type": "object", "required": ["ok", "test_was_passed", "message"], "title": "StartFinishTestSchema"}, "TestPassed": {"properties": {"ok": {"type": "boolean", "title": "Ok"}, "message": {"type": "string", "title": "Message"}, "test_was_passed": {"type": "boolean", "title": "Test Was Passed"}}, "type": "object", "required": ["ok", "message", "test_was_passed"], "title": "TestPassed"}, "UpdateProfile": {"properties": {"first_name": {"type": "string", "title": "First Name"}, "last_name": {"type": "string", "title": "Last Name"}, "middle_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Middle Name"}, "bio": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Bio"}}, "type": "object", "required": ["first_name", "last_name"], "title": "UpdateProfile"}, "UserAnswerSchema": {"properties": {"id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Id"}, "answer_text": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Answer Text"}}, "type": "object", "required": ["id", "answer_text"], "title": "UserAnswerSchema"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}, "src__api_v1__static_test__schemas__ResultTestSchema": {"properties": {"test_id": {"type": "integer", "title": "Test Id"}, "count_correct_answer": {"type": "integer", "title": "Count Correct Answer"}, "count_question": {"type": "integer", "title": "Count Question"}, "score": {"type": "integer", "title": "Score"}, "time_execution": {"type": "integer", "title": "Time Execution"}, "data_answers": {"items": {"$ref": "#/components/schemas/ResultAnswersUserSchema"}, "type": "array", "title": "Data Answers"}}, "type": "object", "required": ["test_id", "count_correct_answer", "count_question", "score", "time_execution", "data_answers"], "title": "ResultTestSchema"}, "src__api_v1__tests__schemas__ResultTestSchema": {"properties": {"test_id": {"type": "integer", "title": "Test Id"}, "count_correct_answer": {"type": "integer", "title": "Count Correct Answer"}, "count_question": {"type": "integer", "title": "Count Question"}, "score": {"type": "integer", "title": "Score"}, "time_execution": {"type": "integer", "title": "Time Execution"}, "data_answers": {"items": {"$ref": "#/components/schemas/ResultAnswersUserSchema"}, "type": "array", "title": "Data Answers"}}, "type": "object", "required": ["test_id", "count_correct_answer", "count_question", "score", "time_execution", "data_answers"], "title": "ResultTestSchema"}}}}